requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Cors==3.0.10
python-dotenv==0.17.1
psycopg2-binary==2.8.6
bcrypt==3.2.0
pyjwt==2.1.0
pytest==6.2.4
pytest-mock==3.6.1
